+++ clear_screen.py
@@ -2,5 +2,4 @@ import os
 from subprocess import call 
 
 def clear(): 
-
-    _ = call('clear' if os.name =='posix' else 'cls') 
+    _ = call('clear' if os.name =='posix' else 'cls')
